+++ Frontend/src/app/services/quotes.service.ts
@@ -13,49 +13,41 @@ export class QuoteService {
 
 	public getQuotes(): Observable<Quote[]> {
 		return this.http.get('http://angularnotes-angularbros.azurewebsites.net/api/Quotes')
-			.map(response => {
-				console.log('response', response.json());
-				return response.json();
-			})
+			.map(response => response.json())
 			.catch(err => err);
 	}
 
 	public getQuote(id: number): Observable<Quote> {
 		return this.http.get(`http://angularnotes-angularbros.azurewebsites.net/api/Quotes/${id}`)
-			.map(response => {
-				return response.json();
-			})
+			.map(response => response.json())
 			.catch(err => err);
 	}
 
 	public getCompanyQuotes(id: any): Observable<any> {
 		return this.http.get(`http://angularnotes-angularbros.azurewebsites.net/api/Quotes?CompanyID=${id}`)
-			.map(response => {
-				return response.json();
-			})
+			.map(response => response.json())
 			.catch(err => err);
 	}
 
-	public createQuote(quote: Quote, companyId?: number): Observable<any> {
+	public createQuote(quote: Quote, companyId: number): Observable<any> {
 		const headers = new Headers({
 			'content-type': 'application/json',
 		});
 		const options = new RequestOptions({headers: headers});
 		return this.http.post(`http://angularnotes-angularbros.azurewebsites.net/api/Quotes?CompanyID=${companyId}`, JSON.stringify(quote), options)
-			.map(res => {
-				return res;
-			})
+			.map(response => response.json())
 			.catch(err => this.handleError(err));
 	}
 
-	public createQuoteLine(quote: Quote, quoteLine: QuoteLine): Observable<any> {
+	public getQuoteLines(quoteId: number): Observable<any> {
 		const headers = new Headers({
 			'content-type': 'application/json',
 		});
 		const options = new RequestOptions({headers: headers});
-		return this.http.put(`http://angularnotes-angularbros.azurewebsites.net/api/QuoteLines/${quote.ID}`, JSON.stringify(quoteLine), options)
-			.map(res => {
-				return res;
+		return this.http.get(`http://angularnotes-angularbros.azurewebsites.net/api/QuoteLines?QuoteID=${quoteId}`)
+			.map(response => {
+				console.log('getQuoteLines', response.json());
+				return response.json();
 			})
 			.catch(err => this.handleError(err));
 	}
@@ -66,37 +58,40 @@ export class QuoteService {
 		});
 		const options = new RequestOptions({headers: headers});
 		return this.http.put(`http://angularnotes-angularbros.azurewebsites.net/api/Quotes/${quote.ID}`, JSON.stringify(quote), options)
-			.map(res => {
-				return res;
-			})
+			.map(response => response)
 			.catch(err => this.handleError(err));
 	}
 
-	public updateQuoteLine(quote: Quote, id: number): Observable<any> {
+	public createQuoteLine(quoteLine: QuoteLine, quote: Quote): Observable<any> {
 		const headers = new Headers({
 			'content-type': 'application/json',
 		});
 		const options = new RequestOptions({headers: headers});
-		return this.http.put(`http://angularnotes-angularbros.azurewebsites.net/api/QutoeLines/${id}`, JSON.stringify(quote), options)
-			.map(res => {
-				return res;
-			})
+		return this.http.post(`http://angularnotes-angularbros.azurewebsites.net/api/QuoteLines?QuoteID=${quote.ID}`, JSON.stringify(quoteLine), options)
+			.map(response => response)
+			.catch(err => this.handleError(err));
+	}
+
+	public updateQuoteLine(quoteLine: QuoteLine): Observable<any> {
+		console.log('quotesLine service Update', quoteLine);
+		const headers = new Headers({
+			'content-type': 'application/json',
+		});
+		const options = new RequestOptions({headers: headers});
+		return this.http.put(`http://angularnotes-angularbros.azurewebsites.net/api/QuoteLines/${quoteLine.ID}`, JSON.stringify(quoteLine), options)
+			.map(response => response)
 			.catch(err => this.handleError(err));
 	}
 
 	public deleteQuote(quoteId: number): Observable<Quote> {
 		return this.http.delete(`http://angularnotes-angularbros.azurewebsites.net/api/Quotes/${quoteId}`)
-			.map(response => {
-				return response.json();
-			})
+			.map(response => response.json())
 			.catch(err => err);
 	}
 
-	public deleteQuoteLine(quoteLine: QuoteLine): Observable<Quote> {
-		return this.http.delete(`http://angularnotes-angularbros.azurewebsites.net/api/Quotes/${quoteLine.ID}`)
-			.map(response => {
-				return response.json();
-			})
+	public deleteQuoteLine(quoteId: number): Observable<Quote> {
+		return this.http.delete(`http://angularnotes-angularbros.azurewebsites.net/api/QuoteLines/${quoteId}`)
+			.map(response => response.json())
 			.catch(err => err);
 	}